finances/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "." }

[project]
name = "finances"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.3",
    "python-dateutil>=2.9.0.post0",
]

[project.scripts]
pie-chart = "expense_analyzer:main"
